agent_test/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
python-dotenv>=1.0.0

# FastAPI
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0

# HTTP 请求
requests>=2.31.0
